hibernate

+++ 18_jpa_hibernate/hibernate_part_two/src/main/java/com/nixsolutions/dao/OrderPartDAO.java
@@ -6,7 +6,7 @@ import com.nixsolutions.hibernate.entity.OrderPart;
 
 public interface OrderPartDAO extends GenericDAO<OrderPart> {
 	
-	public OrderPart getByPK(int orderId, int partId);
+	public OrderPart getByPK(long orderId, int partId);
 	
 	public List<OrderPart> getByOrderId(long orderId);
 }
